ASoC TFA9890: add sysfs interface for battery protection register.
Some users have reported issues with volume at lower battery levels,
and for them to be that noticeable of a drop-off this is the likely cause.

I would suggest just sticking with the batt_protection switching, as the
full BATT_CTL register is still a bit of a mystery. I highly doubt it has
0.1mV precision, so maybe the lower bits mean something else? Unsure, but
it's there to play with for anybody curious.

Signed-off-by: Kyle Repinski <repinski23@gmail.com>